requirements.txt
discord.py==1.3.3
python-dotenv==0.13.0
janome==0.3.10